changing project architecture to clean
