+++ lib/src/core/constant/app_styles/app_styles.dart
@@ -1,0 +1,18 @@
+import 'package:codehub/src/core/constant/app_colors/app_colors.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+const titleStyle =
+    TextStyle(color: textColor, fontWeight: FontWeight.normal, fontSize: 24);
+const titleBoldStyle =
+    TextStyle(color: textColor, fontWeight: FontWeight.bold, fontSize: 24);
+const bodyStyle =
+    TextStyle(color: textColor, fontWeight: FontWeight.normal, fontSize: 19);
+const bodyBoldStyle =
+    TextStyle(color: textColor, fontWeight: FontWeight.bold, fontSize: 19);
+const underlineStyle =
+    TextStyle(color: textColor, fontWeight: FontWeight.normal, fontSize: 15);
+const hintStyle =
+    TextStyle(color: Colors.grey, fontWeight: FontWeight.normal, fontSize: 13);
+const underlineBoldStyle =
+    TextStyle(color: textColor, fontWeight: FontWeight.bold, fontSize: 15);
