+++ lib/src/injector.dart
