+++ lib/src/presentation/blocs/new_snippet_bloc/new_snippet_event.dart
