+++ lib/src/presentation/blocs/new_snippet_bloc/new_snippet_state.dart
