+++ lib/utils/custom_snack_bar/custom_snack_bar.dart
@@ -1,6 +1,5 @@
-import 'package:codehub/constant/app_colors/app_colors.dart';
-import 'package:codehub/constant/app_styles/app_styles.dart';
-import 'package:flutter/cupertino.dart';
+import 'package:codehub/src/core/constant/app_colors/app_colors.dart';
+import 'package:codehub/src/core/constant/app_styles/app_styles.dart';
 import 'package:flutter/material.dart';
 
 customSnackBar(BuildContext context, String title, {bool isSuccess = false}) =>
